requirements.txt
# Requirements
# ===========

# This file specifies the dependencies required to run the application.

# sentence-transformers (version 3.0.1) is required for natural language processing tasks.
sentence-transformers==3.0.1

# pinecone-client is required for interacting with the Pinecone vector database.
pinecone-client

# langchain (version 0.2.5) is required for language model functionality.
langchain==0.2.5

# Flask is required for building the web application.
flask

# pypdf is required for PDF processing.
pypdf

# python-dotenv is required for loading environment variables from a .env file.
python-dotenv

langchain-pinecone

langchain-community

langchain-huggingface

python-box

ensure

pyYAML

types-pyYAML


# The `-e .` flag specifies that the current directory should be treated as a package.
-e .
